Offline sent of reviews on Mobile application now checks for the review survey end date

+++ FeedBackMonkeyMobile/Android/app/src/main/java/cdioil/feedbackmonkey/application/MainMenuActivity.java
@@ -31,6 +31,8 @@ import org.xml.sax.SAXException;
 
 import java.io.File;
 import java.io.IOException;
+import java.util.Calendar;
+import java.util.GregorianCalendar;
 import java.util.List;
 
 import javax.net.ssl.HttpsURLConnection;
@@ -433,30 +435,40 @@ public class MainMenuActivity extends AppCompatActivity {
                     String fileContent = ReviewXMLService.parseReviewFile(f);
 
                     String reviewID = f.getName().split("_")[1];
+                    String surveyEndDate=ReviewXMLService.instance().getSurveyEndDate();
+                    String[] dateSplitted=surveyEndDate.split("-");
+                    String[] timeSplitted=surveyEndDate.split(":");
+                    Calendar endDate=new GregorianCalendar(Integer.parseInt(dateSplitted[0])
+                            ,Integer.parseInt(dateSplitted[1])
+                            ,Integer.parseInt(dateSplitted[2])
+                            ,Integer.parseInt(timeSplitted[0])
+                            ,Integer.parseInt(timeSplitted[1])
+                            ,Integer.parseInt(timeSplitted[2]));
+                    if(Calendar.getInstance().compareTo(endDate)>0) {
+                        String saveReviewURL = BuildConfig.SERVER_URL
+                                .concat(FeedbackMonkeyAPI.getAPIEntryPoint())
+                                .concat(REVIEWS_RESOURCE_PATH)
+                                .concat(SAVE_REVIEW_RESOURCE_PATH)
+                                .concat("/")
+                                .concat(reviewID);
 
-                    String saveReviewURL = BuildConfig.SERVER_URL
-                            .concat(FeedbackMonkeyAPI.getAPIEntryPoint())
-                            .concat(REVIEWS_RESOURCE_PATH)
-                            .concat(SAVE_REVIEW_RESOURCE_PATH)
-                            .concat("/")
-                            .concat(reviewID);
+                        Thread connectionThread = new Thread(() -> {
+                            try {
+                                Response saveReviewResponse = RESTRequest.create(saveReviewURL)
+                                        .withMediaType(RESTRequest.RequestMediaType.XML).withBody(fileContent).POST();
+                                reviewRestResponseCode = saveReviewResponse.code();
+                            } catch (IOException e) {
+                                e.printStackTrace();
+                            }
+                        });
+                        connectionThread.start();
 
-                    Thread connectionThread = new Thread(() -> {
-                        try {
-                            Response saveReviewResponse = RESTRequest.create(saveReviewURL)
-                                    .withMediaType(RESTRequest.RequestMediaType.XML).withBody(fileContent).POST();
-                            reviewRestResponseCode = saveReviewResponse.code();
-                        } catch (IOException e) {
-                            e.printStackTrace();
+                        connectionThread.join();
+                        if (reviewRestResponseCode == HttpsURLConnection.HTTP_OK) {
+                            f.delete();
+                        } else {
+                            return;
                         }
-                    });
-                    connectionThread.start();
-
-                    connectionThread.join();
-                    if (reviewRestResponseCode == HttpsURLConnection.HTTP_OK) {
-                        f.delete();
-                    } else {
-                        return;
                     }
                 }
             } catch (ParserConfigurationException | IOException | SAXException | InterruptedException | TransformerException e) {
